+++ cli/src/commands/backout.rs
@@ -12,6 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+use bstr::ByteVec as _;
 use clap_complete::ArgValueCandidates;
 use itertools::Itertools as _;
 use jj_lib::object_id::ObjectId;
@@ -22,9 +23,13 @@ use crate::cli_util::CommandHelper;
 use crate::cli_util::RevisionArg;
 use crate::command_error::CommandError;
 use crate::complete;
+use crate::formatter::PlainTextFormatter;
 use crate::ui::Ui;
 
-/// Apply the reverse of a revision on top of another revision
+/// Apply the reverse of given revisions on top of another revision
+///
+/// The description of the new revisions can be customized with the
+/// `templates.backout_description` config variable.
 #[derive(clap::Args, Clone, Debug)]
 pub(crate) struct BackoutArgs {
     /// The revision(s) to apply the reverse of
@@ -67,7 +72,6 @@ pub(crate) fn cmd_backout(
         let destination = workspace_command.resolve_single_rev(ui, revision_str)?;
         parents.push(destination);
     }
-    let mut tx = workspace_command.start_transaction();
     let transaction_description = if to_back_out.len() == 1 {
         format!("back out commit {}", to_back_out[0].id().hex())
     } else {
@@ -77,18 +81,31 @@ pub(crate) fn cmd_backout(
             to_back_out.len() - 1
         )
     };
+    let commits_to_back_out_with_new_commit_descriptions = {
+        let template_text = command
+            .settings()
+            .get_string("templates.backout_description")?;
+        let template = workspace_command.parse_commit_template(ui, &template_text)?;
+
+        to_back_out
+            .into_iter()
+            .map(|commit| {
+                let mut output = Vec::new();
+                template
+                    .format(&commit, &mut PlainTextFormatter::new(&mut output))
+                    .expect("write() to vec backed formatter should never fail");
+                // Template output is usually UTF-8, but it can contain file content.
+                let commit_description = output.into_string_lossy();
+                (commit, commit_description)
+            })
+            .collect_vec()
+    };
+    let mut tx = workspace_command.start_transaction();
     let mut new_base_tree = merge_commit_trees(tx.repo(), &parents)?;
-    for commit_to_back_out in to_back_out {
-        let commit_to_back_out_subject = commit_to_back_out
-            .description()
-            .lines()
-            .next()
-            .unwrap_or_default();
-        let new_commit_description = format!(
-            "Back out \"{}\"\n\nThis backs out commit {}.\n",
-            commit_to_back_out_subject,
-            &commit_to_back_out.id().hex()
-        );
+
+    for (commit_to_back_out, new_commit_description) in
+        commits_to_back_out_with_new_commit_descriptions
+    {
         let old_base_tree = commit_to_back_out.parent_tree(tx.repo())?;
         let old_tree = commit_to_back_out.tree()?;
         let new_tree = new_base_tree.merge(&old_tree, &old_base_tree)?;